probeBrailleDisplay() should return an int (rather than a size_t). (ml)
git-svn-id: svn://mielke.cc/main/brltty@6658 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/brl.h
@@ -117,7 +117,7 @@ typedef BrailleResponseResult BrailleResponseHandler (
   const void *packet, size_t size
 );
 
-extern size_t probeBrailleDisplay (
+extern int probeBrailleDisplay (
   BrailleDisplay *brl, unsigned int retryLimit,
   GioEndpoint *endpoint, int inputTimeout,
   BrailleRequestWriter writeRequest,
